move function to util cpp
